sd3.5-large-turbo/requirements.txt
accelerate==1.0.1
bentoml>=1.3.5
diffusers==0.31.0
pillow==11.0.0
protobuf==5.28.3
sentencepiece==0.2.0
torch==2.4.1
transformers==4.46.0
